Refactor SubprocessKeychain
